General cleanup

+++ Classes/DMEnum.h
@@ -1,9 +1,9 @@
 @interface DMEnum : NSObject
 +(NSString*)nameFor:(NSString*)label;
 +(NSArray*)all;
-+(instancetype)at:(int)ordinal;
++(instancetype)at:(NSUInteger)ordinal;
 +(instancetype)for:(id)name;
 +(void)each:(void (^)(id item))block;
 @property(nonatomic, readonly) NSString* name;
-@property(nonatomic, readonly) int ordinal;
+@property(nonatomic, readonly) NSUInteger ordinal;
 @end
